requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
pika==1.3.2
httpx==0.28.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
pytest==8.4.1
pytest-asyncio==1.1.0
python-multipart==0.0.20
python-dotenv==1.1.1
